requirements.txt
flask
python-dotenv
reportlab
google-generativeai
langchain-google-genai
neo4j
pandas
langchain
langchain_community
torch
qdrant-client
Flask-Cors
sentence-transformers
langchain-huggingface
langchain-qdrant
rank_bm25
streamlit
streamlit-option-menu















